Implement mix_columns and use bytearray instead of bytes
mix_columns is a bit tricky to implement, I hope I got it right.
I use look-up tables for some parts of it, those are defined at
the top of the file.

We use bytearray instead of bytes because it is much more logical and
efficient to do crypto in-place in the data rather tha copying every
time the state array. And that's what bytearray is for, and not bytes.