Switch back to bytes instead of bytearray, finish mix_columns
mix_columns was not working because I was updating the state even though
I still needed the old values in the state. This is  why the result
of mix_columns should be allocated and filled rather than put in place.
This led me to realize it will probably be like that for the other
operations, so I switched back to bytes as data type.