Implementation of RSA encryption and decryption
It is apparently working, but the code still needs to be understood
better, there was a fair amount of copy-pasta in the process.

Plus, the generation of a 1024-bits key is fast, but we don't yet know
how long it takes to generate a 2048-bits one. This might be very long.