Implement CBC mode for AES, fix a bug in aes_decrypt
CBC is not parallelizable but that's fine for now since we don't run
encryption in parallel. We can still easily implement another mode if we
decide to.

The bug being fixed is that we used the keys in the wrong order (first
to last instead of last to first) in aes_decrypt.