Specify Upstream/Downstream file sharing tables
There was a problem with the current state of the file sharing protocol.
As it was before, when Z1 has added the bridge in its relay table, it
would not be able to tell if subsequent messages should go to the
tracker or to Z1. This is solved by adding a new, small table in all
nodes, the "Downstream Relay Table".