Add pseudocode for client, tracker and node
This pseudocode is Python-like, and is only aimed to help and guide the
implementation. Do not copy-paste it, as its structure is probably not suited
to fit in the actual code (with Flask and so forth). Especially, there
is a lot of if-else, which can be replaced by path selection and actual
functions, for example