Check for FSID in transfer_upstream
The code is less clean, but this is the only viable way to solve this
simple problem: the field FSID is in the payload, it's not a top field.
Therefore, the node has to try to decode the payload, and if manages to
do that, maybe there is a FSID in the payload. In that case, we can call
transmit_to_bridge. In all other cases, we just forward upstream.

It's kind of like in real TOR: in TOR, nodes try to decrypt the payload,
and hash the result. If the hash matches (like a MD5 sum) then the
payload is plaintext and can be interpreted. Otherwise, relay.